NFC: Split out IfConfigError and the VersionTuple parsing into their own files
