+++ Sources/SwiftIfConfig/IfConfigEvaluation.swift
@@ -2,7 +2,7 @@
 //
 // This source file is part of the Swift.org open source project
 //
-// Copyright (c) 2014 - 2023 Apple Inc. and the Swift project authors
+// Copyright (c) 2014 - 2024 Apple Inc. and the Swift project authors
 // Licensed under Apache License v2.0 with Runtime Library Exception
 //
 // See https://swift.org/LICENSE.txt for license information
@@ -13,231 +13,6 @@ import SwiftDiagnostics
 import SwiftOperators
 import SwiftSyntax
 
-enum IfConfigError: Error, CustomStringConvertible {
-  case unknownExpression(ExprSyntax)
-  case unhandledFunction(name: String, syntax: ExprSyntax)
-  case requiresUnlabeledArgument(name: String, role: String, syntax: ExprSyntax)
-  case unsupportedVersionOperator(name: String, operator: TokenSyntax)
-  case invalidVersionOperand(name: String, syntax: ExprSyntax)
-  case emptyVersionComponent(syntax: ExprSyntax)
-  case compilerVersionOutOfRange(value: Int, upperLimit: Int, syntax: ExprSyntax)
-  case compilerVersionSecondComponentNotWildcard(syntax: ExprSyntax)
-  case compilerVersionTooManyComponents(syntax: ExprSyntax)
-  case canImportMissingModule(syntax: ExprSyntax)
-  case canImportLabel(syntax: ExprSyntax)
-  case canImportTwoParameters(syntax: ExprSyntax)
-  case ignoredTrailingComponents(version: VersionTuple, syntax: ExprSyntax)
-  case integerLiteralCondition(syntax: ExprSyntax, replacement: Bool)
-
-  var description: String {
-    switch self {
-    case .unknownExpression:
-      return "invalid conditional compilation expression"
-
-    case .unhandledFunction(name: let name, syntax: _):
-      return "build configuration cannot handle '\(name)'"
-
-    case .requiresUnlabeledArgument(name: let name, role: let role, syntax: _):
-      return "\(name) requires a single unlabeled argument for the \(role)"
-
-    case .unsupportedVersionOperator(name: let name, operator: let op):
-      return "'\(name)' version check does not support operator '\(op.trimmedDescription)'"
-
-    case .invalidVersionOperand(name: let name, syntax: let version):
-      return "'\(name)' version check has invalid version '\(version.trimmedDescription)'"
-
-    case .emptyVersionComponent(syntax: _):
-      return "found empty version component"
-
-    case .compilerVersionOutOfRange(value: _, upperLimit: let upperLimit, syntax: _):
-      // FIXME: This matches the C++ implementation, but it would be more useful to
-      // provide the actual value as-written and avoid the mathy [0, N] syntax.
-      return "compiler version component out of range: must be in [0, \(upperLimit)]"
-
-    case .compilerVersionSecondComponentNotWildcard(syntax: _):
-      return "the second version component is not used for comparison in legacy compiler versions"
-
-    case .compilerVersionTooManyComponents(syntax: _):
-      return "compiler version must not have more than five components"
-
-    case .canImportMissingModule(syntax: _):
-      return "canImport requires a module name"
-
-    case .canImportLabel(syntax: _):
-      return "2nd parameter of canImport should be labeled as _version or _underlyingVersion"
-
-    case .canImportTwoParameters(syntax: _):
-      return "canImport can take only two parameters"
-
-    case .ignoredTrailingComponents(version: let version, syntax: _):
-      return "trailing components of version '\(version.description)' are ignored"
-
-    case .integerLiteralCondition(syntax: let syntax, replacement: let replacement):
-      return "'\(syntax.trimmedDescription)' is not a valid conditional compilation expression, use '\(replacement)'"
-    }
-  }
-
-  /// Retrieve the syntax node associated with this error.
-  var syntax: Syntax {
-    switch self {
-    case .unknownExpression(let syntax),
-      .unhandledFunction(name: _, syntax: let syntax),
-      .requiresUnlabeledArgument(name: _, role: _, syntax: let syntax),
-      .invalidVersionOperand(name: _, syntax: let syntax),
-      .emptyVersionComponent(syntax: let syntax),
-      .compilerVersionOutOfRange(value: _, upperLimit: _, syntax: let syntax),
-      .compilerVersionTooManyComponents(syntax: let syntax),
-      .compilerVersionSecondComponentNotWildcard(syntax: let syntax),
-      .canImportMissingModule(syntax: let syntax),
-      .canImportLabel(syntax: let syntax),
-      .canImportTwoParameters(syntax: let syntax),
-      .ignoredTrailingComponents(version: _, syntax: let syntax),
-      .integerLiteralCondition(syntax: let syntax, replacement: _):
-      return Syntax(syntax)
-
-    case .unsupportedVersionOperator(name: _, operator: let op):
-      return Syntax(op)
-    }
-  }
-}
-
-extension IfConfigError: DiagnosticMessage {
-  var message: String { description }
-
-  var diagnosticID: MessageID {
-    .init(domain: "SwiftIfConfig", id: "IfConfigError")
-  }
-
-  var severity: SwiftDiagnostics.DiagnosticSeverity {
-    switch self {
-    case .ignoredTrailingComponents: return .warning
-    default: return .error
-    }
-  }
-
-  private struct SimpleFixItMessage: FixItMessage {
-    var message: String
-
-    var fixItID: MessageID {
-      .init(domain: "SwiftIfConfig", id: "IfConfigFixIt")
-    }
-  }
-
-  var asDiagnostic: Diagnostic {
-    // For the integer literal condition we have a Fix-It.
-    if case .integerLiteralCondition(let syntax, let replacement) = self {
-      return Diagnostic(
-        node: syntax,
-        message: self,
-        fixIt: .replace(
-          message: SimpleFixItMessage(
-            message: "replace with Boolean literal '\(replacement)'"
-          ),
-          oldNode: syntax,
-          newNode: BooleanLiteralExprSyntax(
-            literal: .keyword(replacement ? .true : .false)
-          )
-        )
-      )
-    }
-
-    return Diagnostic(node: syntax, message: self)
-  }
-}
-
-extension VersionTuple {
-  /// Parse a compiler build version of the form "5007.*.1.2.3*", which is
-  /// used by an older if configuration form `_compiler_version("...")`.
-  /// - Parameters:
-  ///   - versionString: The version string for the compiler build version that
-  ///   we are parsing.
-  ///   - versionSyntax: The syntax node that contains the version string, used
-  ///   only for diagnostic purposes.
-  fileprivate init(
-    parsingCompilerBuildVersion versionString: String,
-    _ versionSyntax: ExprSyntax
-  ) throws {
-    components = []
-
-    // Version value are separated by periods.
-    let componentStrings = versionString.split(separator: ".")
-
-    /// Record a component after checking its value.
-    func recordComponent(_ value: Int) throws {
-      let limit = components.isEmpty ? 9223371 : 999
-      if value < 0 || value > limit {
-        throw IfConfigError.compilerVersionOutOfRange(value: value, upperLimit: limit, syntax: versionSyntax)
-      }
-
-      components.append(value)
-    }
-
-    // Parse the components into version values.
-    for (index, componentString) in componentStrings.enumerated() {
-      // Check ahead of time for empty version components
-      if componentString.isEmpty {
-        throw IfConfigError.emptyVersionComponent(syntax: versionSyntax)
-      }
-
-      // The second component is always "*", and is never used for comparison.
-      if index == 1 {
-        if componentString != "*" {
-          throw IfConfigError.compilerVersionSecondComponentNotWildcard(syntax: versionSyntax)
-        }
-        try recordComponent(0)
-        continue
-      }
-
-      // Every other component must be an integer value.
-      guard let component = Int(componentString) else {
-        throw IfConfigError.invalidVersionOperand(name: "_compiler_version", syntax: versionSyntax)
-      }
-
-      try recordComponent(component)
-    }
-
-    // Only allowed to specify up to 5 version components.
-    if components.count > 5 {
-      throw IfConfigError.compilerVersionTooManyComponents(syntax: versionSyntax)
-    }
-
-    // In the beginning, '_compiler_version(string-literal)' was designed for a
-    // different version scheme where the major was fairly large and the minor
-    // was ignored; now we use one where the minor is significant and major and
-    // minor match the Swift language version. Specifically, majors 600-1300
-    // were used for Swift 1.0-5.5 (based on clang versions), but then we reset
-    // the numbering based on Swift versions, so 5.6 had major 5. We assume
-    // that majors below 600 use the new scheme and equal/above it use the old
-    // scheme.
-    //
-    // However, we want the string literal variant of '_compiler_version' to
-    // maintain source compatibility with old checks; that means checks for new
-    // versions have to be written so that old compilers will think they represent
-    // newer versions, while new compilers have to interpret old version number
-    // strings in a way that will compare correctly to the new versions compiled
-    // into them.
-    //
-    // To achieve this, modern compilers divide the major by 1000 and overwrite
-    // the wildcard component with the remainder, effectively shifting the last
-    // three digits of the major into the minor, before comparing it to the
-    // compiler version:
-    //
-    //     _compiler_version("5007.*.1.2.3") -> 5.7.1.2.3
-    //     _compiler_version("1300.*.1.2.3") -> 1.300.1.2.3 (smaller than 5.6)
-    //     _compiler_version( "600.*.1.2.3") -> 0.600.1.2.3 (smaller than 5.6)
-    //
-    // So if you want to specify a 5.7.z.a.b version, we ask users to either
-    // write it as 5007.*.z.a.b, or to use the new 'compiler(>= version)'
-    // syntax instead, which does not perform this conversion.
-    if !components.isEmpty {
-      if components.count > 1 {
-        components[1] = components[0] % 1000
-      }
-      components[0] = components[0] / 1000
-    }
-  }
-}
-
 /// Evaluate the condition of an `#if`.
 /// - Parameters:
 ///   - condition: The condition to evaluate, which we assume has already been
@@ -247,7 +22,8 @@ extension VersionTuple {
 ///   - diagnosticHandler: Receives any diagnostics that are produced by the
 ///     evaluation, whether from errors in the source code or produced by the
 ///     build configuration itself.
-/// - Throws: Throws if an error occurs occur during evaluation. The error will
+/// - Throws: Throws if an error occurs occur during evaluation that prevents
+///   this function from forming a valid result. The error will
 ///   also be provided to the diagnostic handler before doing so.
 /// - Returns: A pair of Boolean values. The first describes whether the
 ///   condition holds with the given build configuration. The second whether
